+++ tests/tests/acc_tests/test_user_registration.py
