+++ tests/tests/acc_tests/test_user_view.py
